refactor: refact Resource class

+++ infrastructure/src/test/kotlin/com/lukinhasssss/admin/catalogo/infrastructure/video/DefaultVideoGatewayTest.kt
@@ -812,7 +812,7 @@ class DefaultVideoGatewayTest {
             assertEquals(expectedPage, currentPage)
             assertEquals(expectedPerPage, perPage)
             assertEquals(expectedTotal, total)
-            assertEquals(expectedItemsCount, items.size.toLong())
+            assertEquals(expectedItemsCount, items.size)
             assertEquals(expectedVideo, items[0].title)
         }
     }
@@ -857,7 +857,7 @@ class DefaultVideoGatewayTest {
             assertEquals(expectedPage, currentPage)
             assertEquals(expectedPerPage, perPage)
             assertEquals(expectedTotal, total)
-            assertEquals(expectedItemsCount, items.size.toLong())
+            assertEquals(expectedItemsCount, items.size)
             assertEquals(expectedVideo, items[0].title)
         }
     }
